fix(CA-174): fix timing hacks, extract StubAuthManager, update state names and import paths
