+++ test/features/project_settings/screenshots/project_creation_screen_screenshot_test.dart
@@ -1,99 +1,17 @@
-// ignore_for_file: no_direct_instantiation
-import 'package:construculator/libraries/project/bloc/project_settings_bloc.dart';
+import 'package:construculator/features/project_settings/presentation/bloc/project_settings_bloc/project_settings_bloc.dart';
 import 'package:construculator/features/project_settings/presentation/pages/project_creation_screen.dart';
 import 'package:construculator/features/project_settings/presentation/widgets/project_name_text_field.dart';
 import 'package:construculator/l10n/generated/app_localizations.dart';
-import 'package:construculator/libraries/auth/data/models/auth_credential.dart';
-import 'package:construculator/libraries/auth/data/models/auth_user.dart';
-import 'package:construculator/libraries/auth/data/models/professional_role.dart';
-import 'package:construculator/libraries/auth/domain/types/auth_types.dart';
-import 'package:construculator/libraries/auth/interfaces/auth_manager.dart';
 import 'package:construculator/libraries/project/testing/fake_project_setting_repository.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import '../../../utils/screenshot/font_loader.dart';
+import '../testing/stub_auth_manager.dart';
 
-const String _testUserId = 'test-user-id';
 const Size _screenSize = Size(390, 844);
 const double _pixelRatio = 1.0;
-
-class _StubAuthManager implements AuthManager {
-  const _StubAuthManager();
-
-  @override
-  AuthResult<UserCredential?> getCurrentCredentials() =>
-      AuthResult.success(
-        UserCredential(
-          id: _testUserId,
-          email: 'test@example.com',
-          metadata: const {},
-          createdAt: DateTime(2025, 1, 1),
-        ),
-      );
-
-  @override
-  Future<AuthResult<UserCredential>> loginWithEmail(
-    String email,
-    String password,
-  ) => throw UnimplementedError();
-
-  @override
-  Future<AuthResult<UserCredential>> registerWithEmail(
-    String email,
-    String password,
-  ) => throw UnimplementedError();
-
-  @override
-  Future<AuthResult> sendOtp(String address, OtpReceiver receiver) =>
-      throw UnimplementedError();
-
-  @override
-  Future<AuthResult<UserCredential>> verifyOtp(
-    String address,
-    String otp,
-    OtpReceiver receiver,
-  ) => throw UnimplementedError();
-
-  @override
-  Future<AuthResult<bool>> resetPassword(String email) =>
-      throw UnimplementedError();
-
-  @override
-  Future<AuthResult<bool>> isEmailRegistered(String email) =>
-      throw UnimplementedError();
-
-  @override
-  Future<AuthResult<void>> logout() => throw UnimplementedError();
-
-  @override
-  bool isAuthenticated() => true;
-
-  @override
-  Future<AuthResult<User?>> getUserProfile(String credentialId) =>
-      throw UnimplementedError();
-
-  @override
-  Future<AuthResult<User?>> createUserProfile(User user) =>
-      throw UnimplementedError();
-
-  @override
-  Future<AuthResult<User?>> updateUserProfile(User user) =>
-      throw UnimplementedError();
-
-  @override
-  Future<AuthResult<UserCredential?>> updateUserPassword(String password) =>
-      throw UnimplementedError();
-
-  @override
-  Future<AuthResult<UserCredential?>> updateUserEmail(String email) =>
-      throw UnimplementedError();
-
-  @override
-  Future<AuthResult<List<ProfessionalRole>>> getProfessionalRoles() =>
-      throw UnimplementedError();
-}
 
 void main() {
   TestWidgetsFlutterBinding.ensureInitialized();
@@ -123,7 +41,7 @@ void main() {
         debugShowCheckedModeBanner: false,
         home: BlocProvider.value(
           value: bloc,
-          child: const ProjectCreationScreen(authManager: _StubAuthManager()),
+          child: const ProjectCreationScreen(authManager: StubAuthManager()),
         ),
       );
 
@@ -137,7 +55,7 @@ void main() {
       addTearDown(tester.view.reset);
 
       await tester.pumpWidget(buildScreen());
-      await tester.pump(const Duration(milliseconds: 100));
+      await tester.pumpAndSettle();
 
       await expectLater(
         find.byType(ProjectCreationScreen),
@@ -168,7 +86,7 @@ void main() {
       addTearDown(tester.view.reset);
 
       await tester.pumpWidget(buildScreen(theme: createTestThemeDark()));
-      await tester.pump(const Duration(milliseconds: 100));
+      await tester.pumpAndSettle();
 
       await expectLater(
         find.byType(ProjectCreationScreen),
