+++ test/features/project_settings/widgets/accessibility/project_creation_screen_a11y_test.dart
@@ -1,12 +1,6 @@
-// ignore_for_file: no_direct_instantiation
+import 'package:construculator/features/project_settings/presentation/bloc/project_settings_bloc/project_settings_bloc.dart';
 import 'package:construculator/features/project_settings/presentation/pages/project_creation_screen.dart';
 import 'package:construculator/l10n/generated/app_localizations.dart';
-import 'package:construculator/libraries/auth/data/models/auth_credential.dart';
-import 'package:construculator/libraries/auth/data/models/auth_user.dart';
-import 'package:construculator/libraries/auth/data/models/professional_role.dart';
-import 'package:construculator/libraries/auth/domain/types/auth_types.dart';
-import 'package:construculator/libraries/auth/interfaces/auth_manager.dart';
-import 'package:construculator/libraries/project/bloc/project_settings_bloc.dart';
 import 'package:construculator/libraries/project/testing/fake_project_setting_repository.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
@@ -15,75 +9,7 @@ import 'package:ripplearc_coreui/ripplearc_coreui.dart';
 
 import '../../../../utils/a11y/a11y_guidelines.dart';
 import '../../../../utils/screenshot/font_loader.dart';
-
-const String _testUserId = 'test-user-id';
-
-class _StubAuthManager implements AuthManager {
-  const _StubAuthManager();
-
-  @override
-  AuthResult<UserCredential?> getCurrentCredentials() => AuthResult.success(
-    UserCredential(
-      id: _testUserId,
-      email: 'test@example.com',
-      metadata: const {},
-      createdAt: DateTime(2025, 1, 1),
-    ),
-  );
-
-  @override
-  Future<AuthResult<UserCredential>> loginWithEmail(String email, String password) =>
-      throw UnimplementedError();
-
-  @override
-  Future<AuthResult<UserCredential>> registerWithEmail(String email, String password) =>
-      throw UnimplementedError();
-
-  @override
-  Future<AuthResult> sendOtp(String address, OtpReceiver receiver) =>
-      throw UnimplementedError();
-
-  @override
-  Future<AuthResult<UserCredential>> verifyOtp(
-    String address,
-    String otp,
-    OtpReceiver receiver,
-  ) => throw UnimplementedError();
-
-  @override
-  Future<AuthResult<bool>> resetPassword(String email) => throw UnimplementedError();
-
-  @override
-  Future<AuthResult<bool>> isEmailRegistered(String email) => throw UnimplementedError();
-
-  @override
-  Future<AuthResult<void>> logout() => throw UnimplementedError();
-
-  @override
-  bool isAuthenticated() => true;
-
-  @override
-  Future<AuthResult<User?>> getUserProfile(String credentialId) =>
-      throw UnimplementedError();
-
-  @override
-  Future<AuthResult<User?>> createUserProfile(User user) => throw UnimplementedError();
-
-  @override
-  Future<AuthResult<User?>> updateUserProfile(User user) => throw UnimplementedError();
-
-  @override
-  Future<AuthResult<UserCredential?>> updateUserPassword(String password) =>
-      throw UnimplementedError();
-
-  @override
-  Future<AuthResult<UserCredential?>> updateUserEmail(String email) =>
-      throw UnimplementedError();
-
-  @override
-  Future<AuthResult<List<ProfessionalRole>>> getProfessionalRoles() =>
-      throw UnimplementedError();
-}
+import '../../testing/stub_auth_manager.dart';
 
 void main() {
   late FakeProjectSettingRepository fakeRepository;
@@ -110,7 +36,7 @@ void main() {
         supportedLocales: AppLocalizations.supportedLocales,
         home: BlocProvider.value(
           value: bloc,
-          child: const ProjectCreationScreen(authManager: _StubAuthManager()),
+          child: const ProjectCreationScreen(authManager: StubAuthManager()),
         ),
       );
